requirements.txt
pytest==2.9.0
suds-jurko
pytest_bdd
pillow
selenium
jsonpickle
pony
pymysql


